Commit 8 - Modified Allstocksanalysis 2 to include coe timers
Added code timer to compare how long it takes the program to run and compared 2017 to 2018.  2018 runs slightly faster than 2017 but not enough to be noticeable to the user running the analysis by the sheet.  Note, run times were slightly higher than those noted in the exercise as an example.  0.5664063 seconds for 2018 & 0.5820313seconds for 2017.